+++ reference/java/java8lambdasandstuff/playlambdas/src/com/csoftz/playlambdas/play/domain/Person.java
@@ -1,0 +1,24 @@
+package com.csoftz.playlambdas.play.domain;
+
+public class Person {
+	String firstName;
+	String lastName;
+
+	public Person(String firstName, String lastName) {
+		super();
+		this.firstName = firstName;
+		this.lastName = lastName;
+	}
+
+	public String getFirstName() {
+		return firstName;
+	}
+
+	public String getLastName() {
+		return lastName;
+	}
+
+	
+	
+	
+}