+++ reference/java/java8lambdasandstuff/playlambdas/src/com/csoftz/playlambdas/play/intr/INamed.java
@@ -1,0 +1,9 @@
+package com.csoftz.playlambdas.play.intr;
+
+public interface INamed {
+
+	default String getName() {
+		return getClass().getName() + "_" + hashCode();
+	}
+
+}