+++ reference/java/java8lambdasandstuff/playlambdas/src/com/csoftz/playlambdas/play/intr/IPerson.java
@@ -1,0 +1,9 @@
+package com.csoftz.playlambdas.play.intr;
+
+public interface IPerson {
+	long getId();
+
+	default String getName() {
+		return "John Q. Public";
+	}
+}